楽観ロックの例外をキャッチする処理の実装

+++ app/services/AdminService.java
@@ -15,80 +15,155 @@ import java.util.*;
 
 import javax.persistence.OptimisticLockException;
 
+/**
+ * insert・updateの処理は楽観ロックがかかっている場合には 例外をキャッチして再試行するようにしている。
+ * 
+ * @author masataka.okudera
+ *
+ */
 public class AdminService {
 	/**
-	 *
+	 * 
 	 * @param template
 	 */
-	public void saveTemplate(Template template) {
-		do {
-			try { 
-				template.save();
-				break;
-			} catch (OptimisticLockException e) {
-				Logger.info("処理が競合しました。");
-				Template newTemplate = this.findTemplateById(
-						template.templateId);
-				newTemplate.templateName = template.templateName;
-				newTemplate.templateMessage = template.templateMessage;
-				newTemplate.accessFlag = template.accessFlag;
-				newTemplate.save();
-			}
-		} while(true);
+	public void deleteTemplate(Template template) {
+		try {
+			template.delete();
+		} catch (OptimisticLockException e) {
+			e.printStackTrace();
+			Logger.info("処理が競合しました。");
+			do {
+				template = this.findTemplateById(template.templateId);
+				if (template == null) {
+					break;
+				}
+				try {
+					template.delete();
+					break;
+				} catch (OptimisticLockException e2) {
+					e2.printStackTrace();
+					Logger.info("処理が競合しました。");
+				}
+			} while (true);
+		}
 	}
 
-	public void updateTemplate(Template template){
-		template.update();
-	}
-
-	public void deleteTemplate(Template template){
-		template.delete();
-	}
-
-	public Template findTemplateById(Long id){
+	public Template findTemplateById(Long id) {
 		return Template.findById(id);
 	}
 
-	public List<Template> findTemplateByUser(Long id){
+	public List<Template> findTemplateByUser(Long id) {
 		return Template.find.where().eq("Member_Member_Id", id).findList();
 	}
 
-	public List<Template> findTemplateByUser(Long id, int flg){
-		if(flg == 0){
+	public List<Template> findTemplateByUser(Long id, int flg) {
+		if (flg == 0) {
 			return Template.find.where().eq("Member_Member_Id", id).eq("accessFlag", "0").findList();
-		}else{
+		} else {
 			return Template.find.where().eq("Member_Member_Id", id).eq("accessFlag", "1").findList();
 		}
 	}
 
-	public void deleteMemberWithTemplate(Long id, int flg){
+	/**
+	 * @param id
+	 * @param flg
+	 */
+	public void deleteMemberWithTemplate(Long id, int flg) {
 		List<Template> list = findTemplateByUser(id, flg);
-		for(Template temp : list){
-			if(temp != null){
-				System.out.println(temp.templateName);
+		for (Template temp : list) {
+			if (temp != null) {
 				temp.member = null;
 				temp.accessFlag = 0;
-				temp.update();
+				try {
+					temp.update();
+				} catch (OptimisticLockException e) {
+					e.printStackTrace();
+					Logger.info("処理が競合しました。");
+					do {
+						temp = this.findTemplateById(temp.templateId);
+						if (temp == null) {
+							break;
+						}
+						temp.member = null;
+						temp.accessFlag = 0;
+						try {
+							temp.update();
+							break;
+						} catch (Exception e2) {
+							e2.printStackTrace();
+							Logger.info("処理が競合しました。");
+						}
+					} while (true);
+				}
 			}
 		}
 	}
 
-	public List<Image> findImageByUser(Long id){
+	/**
+	 * 
+	 * @param id
+	 * @return
+	 */
+	public List<Image> findImageByUser(Long id) {
 		return Image.find.where().eq("Member_Member_id", id).findList();
 	}
 
-	public void deleteMemberWithImage(Long id){
+	/**
+	 * 
+	 * @param id
+	 */
+	public void deleteMemberWithImage(Long id) {
 		List<Image> list = findImageByUser(id);
-		for(Image img : list){
-			if(img != null){
+		for (Image img : list) {
+			if (img != null) {
 				img.member = null;
-				img.update();
+				try {
+					img.update();
+				} catch (OptimisticLockException e) {
+					e.printStackTrace();
+					Logger.info("処理が競合しました。");
+					do {
+						img = Image.find.byId(img.imageId);
+						if (img == null) {
+							break;
+						}
+						img.member = null;
+						try {
+							img.update();
+						} catch (OptimisticLockException e2) {
+							e2.printStackTrace();
+							Logger.info("処理が競合しました。");
+						}
+					} while (true);
+				}
 			}
 		}
 	}
 
-	public void deleteImg(Image img){
-		img.delete();
+	/**
+	 * 
+	 * @param img
+	 */
+	public void deleteImg(Image img) {
+		try {
+			img.delete();
+		} catch (OptimisticLockException e) {
+			e.printStackTrace();
+			Logger.info("処理が競合しました。");
+			do {
+				img = Image.find.byId(img.imageId);
+				if (img == null) {
+					break;
+				}
+				try {
+					img.delete();
+					break;
+				} catch (OptimisticLockException e2) {
+					e2.printStackTrace();
+					Logger.info("処理が競合しました。");
+				}
+			} while (true);
+		}
 	}
 
 	/**
@@ -107,9 +182,9 @@ public class AdminService {
 	 */
 	public boolean memberExists(String memberName) {
 		Member member = Member.find.where().eq("memberName", memberName).findUnique();
-		if(member != null) {
+		if (member != null) {
 			return true;
-		}	else	{
+		} else {
 			return false;
 		}
 	}
@@ -129,15 +204,83 @@ public class AdminService {
 	 * @param member
 	 */
 	public void storeMember(Member member) {
-		member.save();
+		try {
+			member.save();
+		} catch (OptimisticLockException e) {
+			e.printStackTrace();
+			Logger.info("処理が競合しました。");
+			do {
+				Member newMember = this.findMemberById(member.memberId);
+				newMember.memberName = member.memberName;
+				newMember.nickName = member.nickName;
+				newMember.password = member.password;
+				newMember.mail = member.mail;
+				newMember.chooser = member.chooser;
+				newMember.lastLogin = member.lastLogin;
+				try {
+					newMember.save();
+					break;
+				} catch (OptimisticLockException e2) {
+					e2.printStackTrace();
+					Logger.info("処理が競合しました。");
+				}
+			} while (true);
+		}
 	}
 
-	public void updateMember(Member member){
-		member.update();
+	/**
+	 * 
+	 * @param member
+	 */
+	public void updateMember(Member member) {
+		try {
+			member.update();
+		} catch (OptimisticLockException e) {
+			e.printStackTrace();
+			Logger.info("処理が競合しました。");
+			do {
+				Member newMember = this.findMemberById(member.memberId);
+				newMember.memberName = member.memberName;
+				newMember.nickName = member.nickName;
+				newMember.password = member.password;
+				newMember.mail = member.mail;
+				newMember.chooser = member.chooser;
+				newMember.lastLogin = member.lastLogin;
+				try {
+					newMember.update();
+					break;
+				} catch (OptimisticLockException e2) {
+					e2.printStackTrace();
+					Logger.info("処理が競合しました。");
+				}
+			} while (true);
+		}
 	}
 
-	public void deleteMember(Member member){
-		member.delete();
+	/**
+	 * 
+	 * @param member
+	 */
+	public void deleteMember(Member member) {
+		try {
+			member.delete();
+		} catch (OptimisticLockException e) {
+			e.printStackTrace();
+			Logger.info("処理が競合しました。");
+			do {
+				member = this.findMemberById(member.memberId);
+				if (member == null) {
+					break;
+				}
+				try {
+					member.delete();
+					break;
+				} catch (OptimisticLockException e2) {
+					e2.printStackTrace();
+					Logger.info("処理が競合しました。");
+				}
+			} while (true);
+		}
 	}
 
 	/**
@@ -168,14 +311,14 @@ public class AdminService {
 		return Chooser.find.byId(chooserId);
 	}
 
-	public Form<Member> addMemberErrors(Form<Member> form, String error, String errorColumn){
+	public Form<Member> addMemberErrors(Form<Member> form, String error, String errorColumn) {
 		List<ValidationError> errors = new ArrayList<ValidationError>();
 		errors.add(new ValidationError(errorColumn, error));
 		form.errors().put(errorColumn, errors);
 		return form;
 	}
 
-	public Form<ModifyPassword> addPasswordErrors(Form<ModifyPassword> form, String error, String errorColumn){
+	public Form<ModifyPassword> addPasswordErrors(Form<ModifyPassword> form, String error, String errorColumn) {
 		List<ValidationError> errors = new ArrayList<ValidationError>();
 		errors.add(new ValidationError(errorColumn, error));
 		form.errors().put(errorColumn, errors);
@@ -188,12 +331,10 @@ public class AdminService {
 	 * @return
 	 */
 	public Image findImageByImageNameAndMemberId(String imageName, Long memberId) {
-		List<Image> images = Image.find.where()
-				.eq("imageName", imageName).eq("member.memberId", memberId)
-				.findList();
-		if(images.size() > 0) {
+		List<Image> images = Image.find.where().eq("imageName", imageName).eq("member.memberId", memberId).findList();
+		if (images.size() > 0) {
 			return images.get(0);
-		}	else	{
+		} else {
 			return null;
 		}
 	}
